Use renderArea of RenderGraph instead of the full window/framebuffer extent when creating the compile-context default viewportstate
Nominally, Views are above RenderGraphs in VSG scene graphs, however when a View that takes up a subset of the window/framebuffer extent
contains individual RenderGraphs, the recorded default renderArea (viewportstate) was incorrect and resulted in the scene being
drawn to the entire window area, but scissored to the render graphs renderArea. This change ensures that the default
viewportstate is consistent with the rendergraph being compiled.